Update task.js
Task done :) 

Few remarks:
Task 1): It was simple and rather easy.
Task 2): I believe having the same code few times to check different stuff is redundant and maybe another loop should be used or maybe switch:cases instead to make it more optimized. Yet I struggled a little bit with it so I send over something that works but it is not ideal per se. 

Thanks for the challange! I had to learn quickly and it was fun :)